Separate OtherAssets from AccountInformation
